# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: 99Placement Recruitment OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requirements:
    get:
      operationId: listRequirements
      tags: [requirements]
      summary: List requirements
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: urgency
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Requirements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequirementWithCompany"
    post:
      operationId: createRequirement
      tags: [requirements]
      summary: Create a requirement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequirementRequest"
      responses:
        "201":
          description: Created requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
  /requirements/{id}:
    get:
      operationId: getRequirement
      tags: [requirements]
      summary: Get requirement with pipeline stats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequirementWithCompany"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRequirement
      tags: [requirements]
      summary: Update requirement
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequirementRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
    delete:
      operationId: deleteRequirement
      tags: [requirements]
      summary: Delete requirement (soft)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── CANDIDATES ──────────────────────────────────────────────────────────────
...
  /pipeline:
    get:
      operationId: getKanbanPipeline
      tags: [pipeline]
      summary: Get kanban pipeline view
      security:
        - bearerAuth: []
      parameters:
        - name: requirementId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Kanban pipeline grouped by stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KanbanPipeline"
  /pipeline/{id}/stage:
    patch:
      operationId: updatePipelineStage
      tags: [pipeline]
      summary: Move candidate to a different stage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageRequest"
      responses:
        "200":
          description: Updated pipeline entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineEntry"

  # ── SCREENING ────────────────────────────────────────────────────────────────
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfferWithCandidate"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "201":
          description: Created offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferWithCandidate"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get offer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferWithCandidate"
  /offers/{id}/status:
    patch:
      operationId: updateOfferStatus
      tags: [offers]
      summary: Update offer status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfferStatusRequest"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferWithCandidate"

  # ── JOINING ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # AUTH ──────────────────────────────────────────────────────────────────────
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, createdAt]
...
    Requirement:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        recruiterId:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        jobType:
          type: string
        urgency:
          type: string
          enum: [critical, high, normal]
        salaryBand:
          type: string
          nullable: true
        jdText:
          type: string
          nullable: true
        openingsCount:
          type: integer
        status:
          type: string
          enum: [open, on_hold, closed, cancelled]
        deadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, companyId, title, location, jobType, urgency, openingsCount, status, createdAt, updatedAt]
    RequirementWithCompany:
      allOf:
        - $ref: "#/components/schemas/Requirement"
        - type: object
          properties:
            company:
              $ref: "#/components/schemas/Company"
            stageCounts:
              type: object
              additionalProperties:
                type: integer
    CreateRequirementRequest:
      type: object
      properties:
        companyId:
          type: string
        title:
          type: string
        location:
          type: string
        jobType:
          type: string
        urgency:
          type: string
          enum: [critical, high, normal]
        salaryBand:
          type: string
        jdText:
          type: string
        openingsCount:
          type: integer
        status:
          type: string
          enum: [open, on_hold, closed, cancelled]
        deadline:
          type: string
          format: date-time
      required: [companyId, title, location]

    # CANDIDATES ────────────────────────────────────────────────────────────────
    Candidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        initials:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
          nullable: true
        currentRole:
          type: string
          nullable: true
        experienceYears:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
        source:
          type: string
          enum: [referral, portal, social, internal, direct]
        currentCtc:
          type: integer
          nullable: true
        expectedCtc:
          type: integer
          nullable: true
        noticeDays:
          type: integer
          nullable: true
        summary:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, source, createdAt]
...
    PipelineEntry:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        requirementId:
          type: string
        stage:
          type: string
          enum: [sourced, screened, assessed, shortlisted, client_interview, offer, joining, post_joining, rejected, dropped]
        assignedRecruiterId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        stageUpdatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, candidateId, requirementId, stage, stageUpdatedAt, createdAt]
    PipelineEntryWithCandidate:
      allOf:
        - $ref: "#/components/schemas/PipelineEntry"
        - type: object
          properties:
            candidate:
              $ref: "#/components/schemas/Candidate"
    KanbanPipeline:
      type: object
      properties:
        sourced:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        screened:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        assessed:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        shortlisted:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        client_interview:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        offer:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        joining:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        post_joining:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        rejected:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
        dropped:
          type: array
          items:
            $ref: "#/components/schemas/PipelineEntryWithCandidate"
      required: [sourced, screened, assessed, shortlisted, client_interview, offer, joining, post_joining, rejected, dropped]
    UpdateStageRequest:
      type: object
      properties:
        stage:
          type: string
          enum: [sourced, screened, assessed, shortlisted, client_interview, offer, joining, post_joining, rejected, dropped]
        notes:
          type: string
      required: [stage]

    # SCREENING ─────────────────────────────────────────────────────────────────
...
    OfferWithCandidate:
      type: object
      properties:
        id:
          type: string
        pipelineId:
          type: string
        offeredCtc:
          type: integer
          nullable: true
        designation:
          type: string
          nullable: true
        joiningDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [drafted, sent, accepted, rejected, revoked]
        sentAt:
          type: string
          format: date-time
          nullable: true
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        candidate:
          $ref: "#/components/schemas/Candidate"
        createdAt:
          type: string
          format: date-time
      required: [id, pipelineId, status, createdAt, candidate]
    CreateOfferRequest:
      type: object
      properties:
        pipelineId:
          type: string
        offeredCtc:
          type: integer
        designation:
          type: string
        joiningDate:
          type: string
          format: date-time
      required: [pipelineId]
    UpdateOfferStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [drafted, sent, accepted, rejected, revoked]
      required: [status]

    # JOINING ───────────────────────────────────────────────────────────────────
...