# lib/api-zod/src/generated/types/requirement.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 99Placement Recruitment OS API
 * OpenAPI spec version: 0.2.0
 */
import type { RequirementStatus } from './requirementStatus';
import type { RequirementUrgency } from './requirementUrgency';

export interface Requirement {
  id: string;
  companyId: string;
  recruiterId?: string | null;
  title: string;
  location: string;
  jobType: string;
  urgency: RequirementUrgency;
  salaryBand?: string | null;
  jdText?: string | null;
  openingsCount: number;
  status: RequirementStatus;
  deadline?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}